consolidate test
